+++ NEWS.md
@@ -420,6 +420,13 @@
 
 * fixes bug in cellchat to work with on disk matrices
 
+# blaseRtools 0.0.0.9171
+
+* updated trace functions
+* added function to import peaks .bed file from SEACR
+* updated bb_plot_trace_peaks to allow filtering.  This may break some rarely used code.
+
+
 # todo
 
 * biaxial cite-seq plots